added animation for div and completed product routing
