+++ client/src/components/RelatedItemsOutfit/Outfit/outfitCard.jsx
@@ -6,7 +6,7 @@ import OutfitAction from '../styledComponents/styledOutfit/outfitAction';
 import ProductContainer from '../styledComponents/styledRelated/productContainer';
 
 const OutfitCard = ({ productInfo, getProduct, deleteOutfit }) => (
-  <ProductContainer to={{ pathname: `/product/${productInfo.id}` }} className="productContainer" onClick={(event) => getProduct(productInfo.id, event)}>
+  <ProductContainer to={{ pathname: `/product/${productInfo.id}` }} className="productContainer" onClick={() => { window.location.reload(); getProduct(productInfo.id); }}>
     <OutfitAction type="button" name={productInfo.styleId} onClick={(event) => { event.preventDefault(); deleteOutfit(event); }}>✖</OutfitAction>
     <ImageGallery photos={productInfo.photos} category={productInfo.category} />
     <Descriptions productInfo={productInfo} />
