+++ client/src/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/app';
-// import App2 from './components/app2';
 
 ReactDOM.render(<App />, document.getElementById('app'));
